Moved EntryType to its own file.
