+++ BibTeXLibrary/BibTeXLibrary.csproj
@@ -50,6 +50,7 @@
     <Compile Include="BibEntry.cs" />
     <Compile Include="BibParser.cs" />
     <Compile Include="Config.cs" />
+    <Compile Include="EntryType.cs" />
     <Compile Include="ParseErrorException.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Token.cs" />
